replaced float with double for all drawing functions and corrected a serious error in the rotation function

+++ src/RcppExports.cpp
@@ -6,22 +6,22 @@
 using namespace Rcpp;
 
 // draw_circle
-DataFrame draw_circle(float centerx, float centery, float centerz, float radius, int resolution);
+DataFrame draw_circle(double centerx, double centery, double centerz, double radius, int resolution);
 RcppExport SEXP recexcavAAR_draw_circle(SEXP centerxSEXP, SEXP centerySEXP, SEXP centerzSEXP, SEXP radiusSEXP, SEXP resolutionSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< float >::type centerx(centerxSEXP);
-    Rcpp::traits::input_parameter< float >::type centery(centerySEXP);
-    Rcpp::traits::input_parameter< float >::type centerz(centerzSEXP);
-    Rcpp::traits::input_parameter< float >::type radius(radiusSEXP);
+    Rcpp::traits::input_parameter< double >::type centerx(centerxSEXP);
+    Rcpp::traits::input_parameter< double >::type centery(centerySEXP);
+    Rcpp::traits::input_parameter< double >::type centerz(centerzSEXP);
+    Rcpp::traits::input_parameter< double >::type radius(radiusSEXP);
     Rcpp::traits::input_parameter< int >::type resolution(resolutionSEXP);
     rcpp_result_gen = Rcpp::wrap(draw_circle(centerx, centery, centerz, radius, resolution));
     return rcpp_result_gen;
 END_RCPP
 }
 // rotate
-DataFrame rotate(NumericVector x, NumericVector y, NumericVector z, float degrx, float degry, float degrz, float pivotx, float pivoty, float pivotz);
+DataFrame rotate(NumericVector x, NumericVector y, NumericVector z, double degrx, double degry, double degrz, double pivotx, double pivoty, double pivotz);
 RcppExport SEXP recexcavAAR_rotate(SEXP xSEXP, SEXP ySEXP, SEXP zSEXP, SEXP degrxSEXP, SEXP degrySEXP, SEXP degrzSEXP, SEXP pivotxSEXP, SEXP pivotySEXP, SEXP pivotzSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -29,26 +29,26 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericVector >::type x(xSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type y(ySEXP);
     Rcpp::traits::input_parameter< NumericVector >::type z(zSEXP);
-    Rcpp::traits::input_parameter< float >::type degrx(degrxSEXP);
-    Rcpp::traits::input_parameter< float >::type degry(degrySEXP);
-    Rcpp::traits::input_parameter< float >::type degrz(degrzSEXP);
-    Rcpp::traits::input_parameter< float >::type pivotx(pivotxSEXP);
-    Rcpp::traits::input_parameter< float >::type pivoty(pivotySEXP);
-    Rcpp::traits::input_parameter< float >::type pivotz(pivotzSEXP);
+    Rcpp::traits::input_parameter< double >::type degrx(degrxSEXP);
+    Rcpp::traits::input_parameter< double >::type degry(degrySEXP);
+    Rcpp::traits::input_parameter< double >::type degrz(degrzSEXP);
+    Rcpp::traits::input_parameter< double >::type pivotx(pivotxSEXP);
+    Rcpp::traits::input_parameter< double >::type pivoty(pivotySEXP);
+    Rcpp::traits::input_parameter< double >::type pivotz(pivotzSEXP);
     rcpp_result_gen = Rcpp::wrap(rotate(x, y, z, degrx, degry, degrz, pivotx, pivoty, pivotz));
     return rcpp_result_gen;
 END_RCPP
 }
 // draw_sphere
-DataFrame draw_sphere(float centerx, float centery, float centerz, float radius, int phires, int thetares);
+DataFrame draw_sphere(double centerx, double centery, double centerz, double radius, int phires, int thetares);
 RcppExport SEXP recexcavAAR_draw_sphere(SEXP centerxSEXP, SEXP centerySEXP, SEXP centerzSEXP, SEXP radiusSEXP, SEXP phiresSEXP, SEXP thetaresSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< float >::type centerx(centerxSEXP);
-    Rcpp::traits::input_parameter< float >::type centery(centerySEXP);
-    Rcpp::traits::input_parameter< float >::type centerz(centerzSEXP);
-    Rcpp::traits::input_parameter< float >::type radius(radiusSEXP);
+    Rcpp::traits::input_parameter< double >::type centerx(centerxSEXP);
+    Rcpp::traits::input_parameter< double >::type centery(centerySEXP);
+    Rcpp::traits::input_parameter< double >::type centerz(centerzSEXP);
+    Rcpp::traits::input_parameter< double >::type radius(radiusSEXP);
     Rcpp::traits::input_parameter< int >::type phires(phiresSEXP);
     Rcpp::traits::input_parameter< int >::type thetares(thetaresSEXP);
     rcpp_result_gen = Rcpp::wrap(draw_sphere(centerx, centery, centerz, radius, phires, thetares));
@@ -56,7 +56,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // scale
-DataFrame scale(NumericVector x, NumericVector y, NumericVector z, float scalex, float scaley, float scalez);
+DataFrame scale(NumericVector x, NumericVector y, NumericVector z, double scalex, double scaley, double scalez);
 RcppExport SEXP recexcavAAR_scale(SEXP xSEXP, SEXP ySEXP, SEXP zSEXP, SEXP scalexSEXP, SEXP scaleySEXP, SEXP scalezSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -64,9 +64,9 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericVector >::type x(xSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type y(ySEXP);
     Rcpp::traits::input_parameter< NumericVector >::type z(zSEXP);
-    Rcpp::traits::input_parameter< float >::type scalex(scalexSEXP);
-    Rcpp::traits::input_parameter< float >::type scaley(scaleySEXP);
-    Rcpp::traits::input_parameter< float >::type scalez(scalezSEXP);
+    Rcpp::traits::input_parameter< double >::type scalex(scalexSEXP);
+    Rcpp::traits::input_parameter< double >::type scaley(scaleySEXP);
+    Rcpp::traits::input_parameter< double >::type scalez(scalezSEXP);
     rcpp_result_gen = Rcpp::wrap(scale(x, y, z, scalex, scaley, scalez));
     return rcpp_result_gen;
 END_RCPP
